refactor: rename Modal to LegacyModal
