+++ frontend/src/lib/modals/LegacyModal.svelte
