+++ frontend/src/lib/modals/canisters/LinkCanisterModal.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
   import AttachCanister from "../../components/canisters/AttachCanister.svelte";
   import { i18n } from "../../stores/i18n";
-  import Modal from "../Modal.svelte";
+  import LegacyModal from "../LegacyModal.svelte";
 </script>
 
-<Modal on:nnsClose size="big">
+<LegacyModal on:nnsClose size="big">
   <svelte:fragment slot="title"
     ><span data-tid="link-canister-modal-title"
       >{$i18n.canisters.attach_canister}</span
@@ -13,7 +13,7 @@
   <section>
     <AttachCanister on:nnsClose />
   </section>
-</Modal>
+</LegacyModal>
 
 <style lang="scss">
   @use "../../themes/mixins/modal";
