+++ frontend/src/lib/modals/neurons/FollowNeuronsModal.svelte
@@ -2,17 +2,17 @@
   import type { NeuronInfo } from "@dfinity/nns";
   import EditFollowNeurons from "../../components/neurons/EditFollowNeurons.svelte";
   import { i18n } from "../../stores/i18n";
-  import Modal from "../Modal.svelte";
+  import LegacyModal from "../LegacyModal.svelte";
 
   export let neuron: NeuronInfo;
 </script>
 
-<Modal size="big" on:nnsClose>
+<LegacyModal size="big" on:nnsClose>
   <span slot="title">{$i18n.neurons.follow_neurons_screen}</span>
   <section>
     <EditFollowNeurons {neuron} />
   </section>
-</Modal>
+</LegacyModal>
 
 <style lang="scss">
   @use "../../themes/mixins/modal";
